Add transactor type and start adding tests
Adding tests to the postgres package for the added functionality.
Currently just a single happy path test is in place, so need to add
some tests to try and hit some of the expected failure conditions
Also in this commit we add a Transactor type to abstract some
Postgres relateld housekeeping tasks.